Feature ETP-1234: Added new TextInput and StringSelector
