+++ packages/MainUI/components/Form/FormView/oldSelectors/StringSelector.tsx
@@ -1,0 +1,3 @@
+import TextInputBase from '@workspaceui/componentlibrary/src/components/Input/TextInput/TextInputBase';
+
+export const StringSelector = TextInputBase;
